+++ src/main/java/noobchain/transaction/transactionoutput/Header.java
@@ -6,7 +6,7 @@ public class Header {
 
     private String id;
     private PublicKey recipient; //also known as the new owner of these coins.
-    private float value; //the amount of coins they own
+    private double value; //the amount of coins they own
 
     public String getId() {
         return id;
@@ -24,11 +24,11 @@ public class Header {
         this.recipient = recipient;
     }
 
-    public float getValue() {
+    public double getValue() {
         return value;
     }
 
-    void setValue(float value) {
+    void setValue(double value) {
         this.value = value;
     }
 }
